+++ 1_Processor/STM32F4/BSPLIB/src/401/can.c
@@ -31,7 +31,7 @@ extern "C" {
 *
 * Return:
 *
-* Cpu_Time:  
+* Cpu_Time:
 *
 * History:
 ***********************************************************************************************************************/
@@ -39,7 +39,7 @@ void HF_CAN_Init(uint8_t CAN_Channel , uint8_t GPIO_AF , uint8_t Mode ,
                  uint8_t T_SJW , uint8_t T_BS1 , uint8_t T_BS2 , uint16_t BRP)
 {
 
-}	
+}
 
 /***********************************************************************************************************************
 * Function:
@@ -52,11 +52,11 @@ void HF_CAN_Init(uint8_t CAN_Channel , uint8_t GPIO_AF , uint8_t Mode ,
 *
 * Return:
 *
-* Cpu_Time:  
+* Cpu_Time:
 *
 * History:
 ***********************************************************************************************************************/
-uint8_t HF_CANTX_Message(uint8_t CAN_Channel , uint8_t StdId  , uint8_t ExtId  ,  uint8_t *TxBuf , uint8_t Length)
+uint8_t HF_CANTX_Message(uint8_t CAN_Channel , uint32_t StdId  , uint32_t ExtId  ,  uint8_t *TxBuf , uint8_t Length)
 {
 
 }
@@ -64,4 +64,3 @@ uint8_t HF_CANTX_Message(uint8_t CAN_Channel , uint8_t StdId  , uint8_t ExtId  ,
 #ifdef __cplusplus
 }
 #endif
-
